requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.5
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
charset-normalizer==3.2.0
    # via aiohttp
discord-py==2.3.2
    # via -r requirements.in
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
pyyaml==6.0.1
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp
